# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for personal code editor with AI analysis and GitHub integration
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Project {
  id: string;
  name: string;
  createdAt: string;
  fileCount: number;
  sizeBytes: number;
}

export type FileNodeType = (typeof FileNodeType)[keyof typeof FileNodeType];

export const FileNodeType = {
  file: "file",
  directory: "directory",
} as const;

export interface FileNode {
  name: string;
  path: string;
  type: FileNodeType;
  children?: FileNode[];
}

export interface ProjectDetail {
  id: string;
  name: string;
  createdAt: string;
  fileCount: number;
  sizeBytes: number;
  tree: FileNode;
}

export interface FileContent {
  path: string;
  content: string;
  language: string;
  isBinary: boolean;
}

export interface AnalyzeFileRequest {
  projectId: string;
  filePath: string;
  content: string;
}

export interface AnalyzeFolderRequest {
  projectId: string;
  folderPath: string;
}

export interface AiAnalysisResult {
  analysis: string;
  model: string;
}

export interface CreateGithubRepoRequest {
  projectId: string;
  repoName: string;
  /** @nullable */
  description?: string | null;
  isPrivate: boolean;
}

export interface CreateGithubRepoResult {
  repoUrl: string;
  repoName: string;
  filesCommitted: number;
}

export interface Settings {
  aiApiKeySet: boolean;
  /** @nullable */
  aiBaseUrl?: string | null;
  /** @nullable */
  aiModel?: string | null;
  githubTokenSet: boolean;
}

export interface UpdateSettingsRequest {
  /** @nullable */
  aiApiKey?: string | null;
  /** @nullable */
  aiBaseUrl?: string | null;
  /** @nullable */
  aiModel?: string | null;
  /** @nullable */
  githubToken?: string | null;
}

export type UploadProjectBody = {
  file: Blob;
  name?: string;
};

export type GetFileContentParams = {
  /**
   * The relative path to the file within the project
   */
  path: string;
};