# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for personal code editor with AI analysis and GitHub integration
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project management (ZIP uploads)
  - name: files
    description: File operations within projects
  - name: ai
    description: AI analysis operations
  - name: github
    description: GitHub integration
  - name: settings
    description: App settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all uploaded projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: uploadProject
      tags: [projects]
      summary: Upload a ZIP file to create a new project
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
              required:
                - file
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details including file tree
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details with file tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{projectId}/files:
    get:
      operationId: getFileContent
      tags: [files]
      summary: Get the content of a specific file within a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: The relative path to the file within the project
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/analyze-file:
    post:
      operationId: analyzeFile
      tags: [ai]
      summary: Analyze a single file using configured AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeFileRequest"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisResult"
        "400":
          description: Bad request or AI not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/analyze-folder:
    post:
      operationId: analyzeFolder
      tags: [ai]
      summary: Analyze a folder's contents using configured AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeFolderRequest"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisResult"
        "400":
          description: Bad request or AI not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /github/create-repo:
    post:
      operationId: createGithubRepo
      tags: [github]
      summary: Create a GitHub repository and push all project files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGithubRepoRequest"
      responses:
        "200":
          description: Repository created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateGithubRepoResult"
        "400":
          description: Bad request or GitHub not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get current settings (without sensitive values)
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
        fileCount:
          type: integer
        sizeBytes:
          type: integer
      required:
        - id
        - name
        - createdAt
        - fileCount
        - sizeBytes

    FileNode:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        children:
          type: array
          items:
            $ref: "#/components/schemas/FileNode"
      required:
        - name
        - path
        - type

    ProjectDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
        fileCount:
          type: integer
        sizeBytes:
          type: integer
        tree:
          $ref: "#/components/schemas/FileNode"
      required:
        - id
        - name
        - createdAt
        - fileCount
        - sizeBytes
        - tree

    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        language:
          type: string
        isBinary:
          type: boolean
      required:
        - path
        - content
        - language
        - isBinary

    AnalyzeFileRequest:
      type: object
      properties:
        projectId:
          type: string
        filePath:
          type: string
        content:
          type: string
      required:
        - projectId
        - filePath
        - content

    AnalyzeFolderRequest:
      type: object
      properties:
        projectId:
          type: string
        folderPath:
          type: string
      required:
        - projectId
        - folderPath

    AiAnalysisResult:
      type: object
      properties:
        analysis:
          type: string
        model:
          type: string
      required:
        - analysis
        - model

    CreateGithubRepoRequest:
      type: object
      properties:
        projectId:
          type: string
        repoName:
          type: string
        description:
          type: ["string", "null"]
        isPrivate:
          type: boolean
      required:
        - projectId
        - repoName
        - isPrivate

    CreateGithubRepoResult:
      type: object
      properties:
        repoUrl:
          type: string
        repoName:
          type: string
        filesCommitted:
          type: integer
      required:
        - repoUrl
        - repoName
        - filesCommitted

    Settings:
      type: object
      properties:
        aiApiKeySet:
          type: boolean
        aiBaseUrl:
          type: ["string", "null"]
        aiModel:
          type: ["string", "null"]
        githubTokenSet:
          type: boolean
      required:
        - aiApiKeySet
        - githubTokenSet

    UpdateSettingsRequest:
      type: object
      properties:
        aiApiKey:
          type: ["string", "null"]
        aiBaseUrl:
          type: ["string", "null"]
        aiModel:
          type: ["string", "null"]
        githubToken:
          type: ["string", "null"]